Add new column for mutation position in sub-peptide
This commit also includes an updated algorithm to
find the corresponding wt epitope sequence for a
mt epitope.

Previously we used a brute force approach and
compared all wt epitope sequences to the mt epitope
and simply picked the one with the most number of
matching amino acids.

Now we use the indel length to find the positions
of wt epitopes to compare.