Fix IEDB Class II bug
Sometimes the start and end positions of the class II IEDB output
refers to the core peptide, and sometimes it refers to the peptide
(IEDB seems to switch this back and forth).

With this change we try to be smarter about how to calculate the
start position of the peptide.