Use eptiope_length to determine spacer wingspan length
When creating the fasta file with all possible peptide-spacer-peptide
combinations we were previously always using a wingspan of 10, which can
result in epitopes that don't overlap the spacer for epitope lengths <=
10.

With this change we now write multiple fasta files, one for epitope
length chosen by the user and use the appropriate wingspan length for
each of them.

When calling IEDB we then use the fasta file that matches the eptiope
length for the specific call to IEDB and its epitope length.