Fix issue with how AGFusion fusion positions are determined
When running AGFusion with the --noncanonical flag, the exons.csv file
will contain exon postions for all possible transcript combinations.
The transcripts weren't being taken into account when parsing this file
to determine the fusion positions. This commit fixes this issue by
looking up the positions for the specific transcripts of the record
currently being parsed.