Update top_score_metric2 to take a list of metrics to prioritize
The rank of all specified metrics will be calculated and the rank
sum will be used to select the best peptide and sort the result files.

This commit also moves the sort_table logic for UpdateTiers the
sort helper module which is now used in both the UpdateTiers and
TopScoreFilter modules for consistency.